// src/app/models/DentistSignupStage2.ts
export interface DentistSignupStage2 {
  dentistUserId: string;
  licenceNumber: string;
}

export interface OfficeSchema {
  name: string;
  address: string;
  city: string;
  zip: string;
  openHours: string;
  officePhoneNumber: string;
  officePhoneNumberType: string;
  email: string;
  officeEmail: string;
}